Throw Exceptions for unset options
In order to not need to check if all necessary command line parameters were given, exceptions are thrown when an unset option is requested. The main program lives in a giant try and exits with a meaningful error messages, if such a case occurs.

Quick and dirty, but it works for this modest use case.